+++ src/main/java/de/sist/gitlab/pipelinemonitor/ui/UntrackedRemoteNotificationState.java
@@ -1,0 +1,11 @@
+package de.sist.gitlab.pipelinemonitor.ui;
+
+import com.intellij.util.messages.Topic;
+
+public interface UntrackedRemoteNotificationState {
+
+    Topic<UntrackedRemoteNotificationState> UNTRACKED_REMOTE_FOUND = Topic.create("Untracked remote found", UntrackedRemoteNotificationState.class);
+
+    void handle(boolean isOpen);
+
+}